backend_core/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.1.0
websockets>=12.0
aiohttp>=3.9.0
colorama>=0.4.6
